Consolidate cryptography client model headers into a single header.
