+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/cryptography/encrypt_parameters.hpp
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include "azure/keyvault/keys/cryptography/encryption_algorithm.hpp"
+#include "azure/keyvault/keys/cryptography/cryptography_client_models.hpp"
 
 #include <string>
 #include <vector>
