+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keyvault_keys.hpp
@@ -9,22 +9,14 @@
 #pragma once
 
 #include "azure/keyvault/keys/cryptography/cryptography_client.hpp"
+#include "azure/keyvault/keys/cryptography/cryptography_client_models.hpp"
 #include "azure/keyvault/keys/cryptography/cryptography_client_options.hpp"
 #include "azure/keyvault/keys/cryptography/decrypt_parameters.hpp"
-#include "azure/keyvault/keys/cryptography/decrypt_result.hpp"
 #include "azure/keyvault/keys/cryptography/encrypt_parameters.hpp"
-#include "azure/keyvault/keys/cryptography/encrypt_result.hpp"
-#include "azure/keyvault/keys/cryptography/encryption_algorithm.hpp"
-#include "azure/keyvault/keys/cryptography/key_wrap_algorithm.hpp"
-#include "azure/keyvault/keys/cryptography/sign_result.hpp"
-#include "azure/keyvault/keys/cryptography/signature_algorithm.hpp"
-#include "azure/keyvault/keys/cryptography/unwrap_result.hpp"
-#include "azure/keyvault/keys/cryptography/verify_result.hpp"
-#include "azure/keyvault/keys/cryptography/wrap_result.hpp"
 #include "azure/keyvault/keys/delete_key_operation.hpp"
 #include "azure/keyvault/keys/dll_import_export.hpp"
 #include "azure/keyvault/keys/key_client.hpp"
-#include "azure/keyvault/keys/key_client_options.hpp"
 #include "azure/keyvault/keys/key_client_models.hpp"
+#include "azure/keyvault/keys/key_client_options.hpp"
 #include "azure/keyvault/keys/list_keys_responses.hpp"
 #include "azure/keyvault/keys/recover_deleted_key_operation.hpp"
