+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/encrypt_result.cpp
@@ -6,7 +6,7 @@
 
 #include "../private/cryptography_serializers.hpp"
 #include "../private/key_constants.hpp"
-#include "azure/keyvault/keys/cryptography/encrypt_result.hpp"
+#include "azure/keyvault/keys/cryptography/cryptography_client_models.hpp"
 
 #include <string>
 #include <vector>
