+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/encryption_algorithm.cpp
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/keyvault/keys/cryptography/encryption_algorithm.hpp"
 #include "../private/key_constants.hpp"
+#include "azure/keyvault/keys/cryptography/cryptography_client_models.hpp"
 
 namespace Azure {
   namespace Security {
