+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/key_wrap_parameters.cpp
@@ -5,7 +5,7 @@
 
 #include "../private/cryptography_serializers.hpp"
 #include "../private/key_constants.hpp"
-#include "azure/keyvault/keys/cryptography/encrypt_parameters.hpp"
+#include "azure/keyvault/keys/cryptography/cryptography_client_models.hpp"
 
 #include <string>
 #include <vector>
