+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/verify_result.cpp
@@ -5,7 +5,7 @@
 
 #include "../private/cryptography_serializers.hpp"
 #include "../private/key_constants.hpp"
-#include "azure/keyvault/keys/cryptography/verify_result.hpp"
+#include "azure/keyvault/keys/cryptography/cryptography_client_models.hpp"
 
 #include <string>
 #include <vector>
